requirements.txt
langgraph>=0.2.60
langchain>=0.3.15
langchain-openai>=0.2.14
langchain-experimental>=0.3.3  # PythonAstREPLTool
pandas>=2.2.0
openpyxl>=3.1.5                # .xlsx support
xlrd>=2.0.1                    # Legacy .xls support
python-dotenv>=1.0.1
tabulate>=0.9.0                # Pretty-print tables in terminal
rich>=13.0.0                   # Improved terminal output (colors, panels)
